[C--] propagate location through C--
Change the type of C-- nodes to be a record holding the location information and the node description (of C-- type).

This was the nig missing piece for location propagation, So now, location information goes from the front-end to the linear code.

The remaining bit is to emit the location information in the assembly.